OUTPUT_DOCUMENTS/GOOGLE GEMINI/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
